+++ BankAccounts/Bank.hpp
@@ -1,7 +1,32 @@
+// ? imports
 #include <vector>
 #include <string>
+
+// ? headerfiles
 #include "Account.hpp"
 
+/**
+ * ? @brief Bank
+ * * stores the accounts of users in vector
+ * * creates bank accounts and returns bank accounts
+ * * get single account with the help of Account::code (string)
+ *
+ * ? @attributes
+ * * accounts vector of type Account class
+ *
+ * ? @functions
+ * * createBankAccount(std::string, std::string, int):
+ * *   - create Account object with code, title and initial ammount
+ * *   - and adds it in the accounts vector
+ *
+ * * getAllAccounts():
+ * *   - returns accounts vector
+ *
+ * * getAccount(std::string):
+ * *   - loops on the accounts vector and
+ * *   - returns single account with the help of Account::code
+ *
+ */
 class Bank
 {
 private:
@@ -13,7 +38,7 @@ public:
         accounts = {};
     }
 
-    void createBankAccount(std::string title, std::string code, double initialDeposit)
+    void createBankAccount(std::string title, std::string code, int initialDeposit)
     {
         Account account(title, code, initialDeposit);
         accounts.push_back(account);
